+++ frontend/cypress/support/e2e.js
@@ -14,7 +14,10 @@
 // ***********************************************************
 
 // Import commands.js using ES2015 syntax:
-import './commands';
+import './commands/loginCommands';
+import './commands/sharedCommands';
+import './commands/workerCommands';
+import './commands/workplaceCommands';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
